Support literal type in OpenAPI
