+++ dmr/openapi/mappers/types.py
@@ -1,4 +1,3 @@
-import contextlib
 from collections import OrderedDict, defaultdict, deque
 from collections.abc import (
     Callable,
@@ -28,6 +27,7 @@ from typing import (  # noqa: WPS235
     Annotated,
     Any,
     ClassVar,
+    Literal,
     Optional,
     TypeAlias,
     Union,
@@ -44,7 +44,6 @@ from dmr.openapi.objects.schema import Schema
 
 if TYPE_CHECKING:
     from dmr.openapi.generators.schema import SchemaGenerator
-    from dmr.openapi.types import KwargDefinition
 
 
 SchemaPredicate: TypeAlias = Callable[[Any, Any, tuple[Any, ...]], bool]
@@ -58,71 +57,6 @@ ReferencedSchemaCallback: TypeAlias = Callable[
 ]
 SchemaOrCallback: TypeAlias = Schema | SchemaCallback
 _PredicatePair: TypeAlias = tuple[SchemaPredicate, ReferencedSchemaCallback]
-
-
-class KwargMapper:
-    """
-    Class for mapping ``KwargDefinition`` to OpenAPI ``Schema``.
-
-    This class is responsible for converting model-specific constraints
-    into OpenAPI-compliant schema attributes.
-    """
-
-    __slots__ = ()
-
-    # Public API:
-    mapping: ClassVar[Mapping[str, str]] = {
-        'content_encoding': 'content_encoding',
-        'default': 'default',
-        'title': 'title',
-        'description': 'description',
-        'const': 'const',
-        'gt': 'exclusive_minimum',
-        'ge': 'minimum',
-        'lt': 'exclusive_maximum',
-        'le': 'maximum',
-        'multiple_of': 'multiple_of',
-        'min_items': 'min_items',
-        'max_items': 'max_items',
-        'min_length': 'min_length',
-        'max_length': 'max_length',
-        'pattern': 'pattern',
-        'format': 'format',
-        'enum': 'enum',
-        'read_only': 'read_only',
-        'examples': 'examples',
-        'external_docs': 'external_docs',
-    }
-
-    def __call__(
-        self,
-        schema: 'Schema',
-        kwarg_definition: 'KwargDefinition',
-    ) -> dict[str, Any]:
-        """
-        Extract updates for the schema from the kwarg definition.
-
-        Args:
-            schema: The schema object to be updated.
-            kwarg_definition: Data container with constraints.
-
-        """
-        updates: dict[str, Any] = {}
-        for kwarg_field, schema_field in self.mapping.items():
-            kwarg_value = getattr(kwarg_definition, kwarg_field)
-            if kwarg_value is None:
-                continue
-
-            if kwarg_field == 'format':
-                with contextlib.suppress(ValueError):
-                    kwarg_value = OpenAPIFormat(kwarg_value)
-
-            updates[schema_field] = kwarg_value
-
-        if kwarg_definition.schema_extra:
-            updates.update(kwarg_definition.schema_extra)
-
-        return updates
 
 
 def _generic_array(
@@ -156,10 +90,7 @@ def _handle_union(
 ) -> Schema | Reference:
     if not type_args:
         # We know that NoneType is registered in TypeMapper
-        schema = TypeMapper.get_schema(None, generator)
-        # for mypy: we know that `None` is registered
-        assert schema is not None  # noqa: S101
-        return schema
+        return generator(None)
     return Schema(one_of=[generator(type_arg) for type_arg in type_args])
 
 
@@ -178,6 +109,17 @@ def _handle_annotated(
     # TODO: support metadata from `Annotated` to pass to schema's kwargs?
     # Example: `Annotated[int, Schema(maximum=0)]`
     return generator(type_args[0])
+
+
+def _handle_literal(
+    annotation: Any,
+    origin: Any,
+    type_args: tuple[Any, ...],
+    generator: 'SchemaGenerator',
+) -> Schema | Reference:
+    if not type_args:
+        return generator(Any)
+    return Schema(enum=type_args)
 
 
 class TypeMapper:
@@ -265,6 +207,11 @@ class TypeMapper:
         (
             lambda annotation, origin, type_args: origin is Annotated,
             _handle_annotated,
+        ),
+        # `Literal[T, V]`:
+        (
+            lambda annotation, origin, type_args: origin is Literal,
+            _handle_literal,
         ),
     ]
